requirements.txt
opencv-python
FER
pandas
tensorflow
jinja2
aiohttp-jinja2
//flask
//Flask-SocketIO==5.1.2
//python-engineio==4.3.2
//python-socketio==5.6.0
//flask_cors

numpy
keras
Pillow
aiohttp
matplotlib
python-socketio
